% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_kendallB.R
\name{cor_kendallB}
\alias{cor_kendallB}
\title{cor_kendallB}
\usage{
cor_kendallB(x, y)
}
\arguments{
\item{x, y}{input value}
}
\value{
the Kendall’s tau-b correlation between x and y.
}
\description{
Calculate the Kendall’s tau-b correlation between x and y.
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9)
y<-c(2,5,87,3,7,5,3,7,79)
cor_kendallB(x,y)

cor_kendallB(mtcars$mpg,mtcars$cyl)

}
