% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_pearson.R
\name{cor_pearson}
\alias{cor_pearson}
\title{cor_pearson}
\usage{
cor_pearson(x, y)
}
\arguments{
\item{x, y}{input value}
}
\value{
the Pearson correlation between x and y.
}
\description{
Calculate the Pearson correlation between x and y.
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9)
y<-c(2,5,87,3,7,5,3,7,79)
cor_pearson(x,y)

}
