% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_spearman.R
\name{cor_spearman}
\alias{cor_spearman}
\title{cor_spearman}
\usage{
cor_spearman(x, y)
}
\arguments{
\item{x, y}{input value}
}
\value{
the Spearman correlation between x and y.
}
\description{
Calculate the Spearman correlation between x and y.
}
\examples{
x<-c(1,2,3,4,5,6,7,8,9)
y<-c(2,5,87,3,7,5,3,7,79)
cor_spearman(x,y)

cor_spearman(mtcars$mpg,mtcars$cyl)

}
