% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_check.R
\name{cor_check}
\alias{cor_check}
\title{cor_check}
\usage{
cor_check(x, y)
}
\arguments{
\item{x, y}{input value}
}
\value{
the results of test
}
\description{
Check if there is NA, empty cells, or non-numeric values in each variable.
}
\examples{
x<-c(1,2,3,"4","",6)
y<-c(5,3,7,1)
cor_check(x,ycor)

}
